+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [unreleased]
 - **Added** `path` property to error object
 - **Fixed** repeated errors in nested model scenario
+- **Breaking changes**
+    - **Changed** in locale strings {field} with {path}
 
 [3.3.1] - 2017-12-28
 - **Added** new type `any`